Add changes from review
