+++ remote_challenge_evaluation/remote_evaluation_util.py
@@ -1,17 +1,25 @@
+from eval_ai_interface import EvalAI_Interface
+
 
 class RemoteEvaluationUtil:
-    def __init__(self):  
+    def __init__(self):
         self.auth_token = ""  # Go to EvalAI UI to fetch your auth token
         self.evalai_api_server = ""  # For staging server, use -- https://staging.eval.ai; For production server, use -- https://eval.ai
-        self.queue_name = ""  # Please email EvalAI admin (team@cloudcv.org) to get the queue name
-        self.challenge_pk = ""  # Please email EvalAI admin (team@cloudcv.org) to get the challenge primary key
+        self.queue_name = ""  # Check Manage Tab of challenge for queue name
+        self.challenge_pk = ""  # Check Manage Tab of challenge for challenge PK
 
-    def download_and_evaluate(self, submission, evalai, challenge_pk, phase_pk, submission_pk):
-        
+        # Create evalai object
+        self.evalai = EvalAI_Interface(
+            self.auth_token, self.evalai_api_server, self.queue_name, self.challenge_pk
+        )
+
+    def download_and_evaluate(
+        self, submission, challenge_pk, phase_pk, submission_pk
+    ):
 
         # Set the status to running
         # self._update_running(submission, job_name="")
-        
+
         # Download the input file
         # The submission file URL can be found at submission_input_file_url = submission.input_file.url
         # Run the submission with the input file using your own code and data.
@@ -25,12 +33,17 @@ class RemoteEvaluationUtil:
         # )
         pass
 
+    def _update_running(self, submission, job_name):
+        status_data = {
+            "submission": submission,
+            "job_name": job_name,
+            "submission_status": "RUNNING",
+        }
+        update_status = self.evalai.update_submission_status(status_data)
 
-    def _update_running(self, evalai, submission, job_name):
-        status_data = {"submission": submission, "job_name": job_name, "submission_status": "RUNNING"}
-        update_status = evalai.update_submission_status(status_data)
-
-    def _update_failed(self, evalai, phase_pk, submission_pk, submission_error, stdout="", metadata=""):
+    def _update_failed(
+        self, phase_pk, submission_pk, submission_error, stdout="", metadata=""
+    ):
         submission_data = {
             "challenge_phase": phase_pk,
             "submission": submission_pk,
@@ -39,10 +52,17 @@ class RemoteEvaluationUtil:
             "submission_status": "FAILED",
             "metadata": metadata,
         }
-        update_data = evalai.update_submission_data(submission_data)
+        update_data = self.evalai.update_submission_data(submission_data)
 
-
-    def _update_finished(self, evalai, phase_pk, submission_pk, result, submission_error="", stdout="", metadata=""):
+    def _update_finished(
+        self,
+        phase_pk,
+        submission_pk,
+        result,
+        submission_error="",
+        stdout="",
+        metadata="",
+    ):
         submission_data = {
             "challenge_phase": phase_pk,
             "submission": submission_pk,
@@ -52,7 +72,4 @@ class RemoteEvaluationUtil:
             "result": result,
             "metadata": metadata,
         }
-        update_data = evalai.update_submission_data(submission_data)
-
-
-remote_evaluation_util = RemoteEvaluationUtil()
+        update_data = self.evalai.update_submission_data(submission_data)
